Update 3
Added 5 projects to the portfolio section.  Added working links to the live project websites.  Added links to the github pages.  Added headers and styles.  Improved spacing on images and sizes when responding to responsive layouts.  Added technologies used section to the project panes.  Added dummy text and captions.  Added highlight project pane that is twice the width and at the top of the portfolio section.