Update 6
Added education image to education section.  Updated responsiveness and spacing of the image.  Updated spacing of skills section and modified the design to have two columns in one row.  Removed the nav from top and made it constant floating with the ability to open and close in mobile.  Updated summary text in banner and contact section.  Cleaned up unnecessary code.  Adjusted margin and padding in contact section, skills section and banner.